+++ changelog.md
@@ -1,6 +1,9 @@
 Added:
-1.components
-2.loader for button and selection for input text
+
+1. components
+2. loader for button and selection for input text 10. react-datepicker added
+3. react-datepicker added
+4. react-select added
 
 Need to consider: